Refactor admin invitation form
